secure API endpoint

+++ src/Controller/DocumentController.php
@@ -14,15 +14,27 @@ use App\Service\FileUploader;
 
 class DocumentController extends AbstractController
 {
-    public function __construct(private ManagerRegistry $doctrine) {}
+    public function __construct(private ManagerRegistry $doctrine)
+    {
+        try{
+            $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        }catch (\Exception $e){
+            return $this->json('Access denied', 400);
+        }
+    }
 
     #[Route('/api/document', name: 'get_document', methods: ['GET'])]
     public function index(ManagerRegistry $doctrine): JsonResponse
     {
+        $user = $this->getUser();
         $entityManager = $doctrine->getManager();
 
-        $documents = $entityManager->getRepository(Document::class)->findAll();
-
+        if($user->getRole() == 'ROLE_ADMIN') {
+            $documents = $entityManager->getRepository(Document::class)->findAll();
+        }else{
+            $documents = $entityManager->getRepository(Document::class)->findBy(['user' => $user]);
+        }
+      
         $data = [];
 
         foreach ($documents as $document) {
@@ -59,7 +71,7 @@ class DocumentController extends AbstractController
         $document->setCreatedAt(new \DateTime());
 
         $uploadedFile = $request->files->get('file');
-        if($uploadedFile){
+        if ($uploadedFile) {
             $mediaObject = new DocumentMedia();
             $mediaObject->filePath = $fileUploader->upload($uploadedFile);
             $document->setMedia($mediaObject);
@@ -94,6 +106,13 @@ class DocumentController extends AbstractController
             return $this->json('No document found for id' . $id, 404);
         }
 
+        $user = $this->getUser();
+
+        if($user->getId() !== $document->getUser()->getId()){
+            return $this->json('Access denied', 400);
+        }
+
+
         $data = [
             'id' => $document->getId(),
             'title' => $document->getTitle(),
@@ -124,6 +143,12 @@ class DocumentController extends AbstractController
             return $this->json('No document type found for id' . $request->request->get('documentType'), 404);
         }
 
+        $user = $this->getUser();
+
+        if($user->getId() !== $document->getUser()->getId()){
+            return $this->json('Access denied', 400);
+        }
+
         $document->setTitle($request->request->get('title'));
         $document->setDescription($request->request->get('description'));
         $document->setDocumentType($documentType);
@@ -149,6 +174,12 @@ class DocumentController extends AbstractController
 
         if (!$document) {
             return $this->json('No document found for id' . $id, 404);
+        }
+
+        $user = $this->getUser();
+
+        if($user->getId() !== $document->getUser()->getId()){
+            return $this->json('Access denied', 400);
         }
 
         $entityManager->remove($document);
